fix; add streetViewer.vue
